app/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
motor==3.6.0
pymongo==4.9.1
opencv-python-headless==4.10.0.84
numpy==2.0.2
requests==2.32.3
onnxruntime==1.19.2
